backend/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
pillow==9.4.0
django-filter==22.1
MarkupSafe==2.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.2.2
djoser
idna==3.4
itypes==1.2.0
jinja2==3.1.2
oauthlib==3.2.2
pycparser==2.21
pyjwt==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
uritemplate==4.1.1
urllib3==1.26.14
psycopg2-binary==2.9.5
python-dotenv==0.21.1
django-extra-fields==3.0.2
gunicorn==20.1.0